requirements.txt
fastapi
uvicorn
requests
sqlalchemy
apscheduler
pyyaml
